refactor: extract useDataCollection and useData from datacollection
