+++ packages/react/src/experimental/Forms/Fields/Select/__stories__/index.stories.tsx
@@ -3,7 +3,7 @@ import { fn } from "storybook/test"
 import { Select, SelectProps } from "../index"
 
 import { IconType } from "@/components/Utilities/Icon"
-import { createDataSourceDefinition } from "@/experimental/OneDataCollection/useDataSource"
+import { createDataSourceDefinition } from "@/hooks/datasource/useDataSource/useDataSource"
 import { Appearance, Circle, Desktop, Plus } from "@/icons/app"
 import {
   DEPARTMENTS_MOCK,
