+++ packages/react/src/experimental/Forms/Fields/Select/index.tsx
@@ -1,24 +1,27 @@
 import { Icon } from "@/components/Utilities/Icon"
 import {
+  ItemActionsDefinition,
+  NavigationFiltersDefinition,
+  SummariesDefinition,
+} from "@/experimental/exports"
+import { RawTag } from "@/experimental/Information/Tags/RawTag"
+import { GroupHeader } from "@/experimental/OneDataCollection/components/GroupHeader"
+import {
   BaseFetchOptions,
   BaseResponse,
   DataSourceDefinition,
   FiltersDefinition,
+  getDataSourcePaginationType,
   GroupingDefinition,
-  ItemActionsDefinition,
-  NavigationFiltersDefinition,
   PaginatedDataAdapter,
   PromiseOrObservable,
   RecordType,
   SortingsDefinition,
-  SummariesDefinition,
+  useData,
   useDataSource,
-} from "@/experimental/exports"
-import { RawTag } from "@/experimental/Information/Tags/RawTag"
-import { GroupHeader } from "@/experimental/OneDataCollection/components/GroupHeader"
-import { useData, WithGroupId } from "@/experimental/OneDataCollection/useData"
-import { getDataSourcePaginationType } from "@/experimental/OneDataCollection/useDataSource"
-import { useGroups } from "@/experimental/OneDataCollection/useGroups"
+  WithGroupId,
+} from "@/hooks/datasource"
+import { useGroups } from "@/hooks/datasource/useGroups"
 import { ChevronDown } from "@/icons/app"
 import { cn } from "@/lib/utils"
 import { InputField, InputFieldProps } from "@/ui/InputField"
