+++ packages/react/src/experimental/Forms/Fields/Select/types.ts
@@ -1,5 +1,5 @@
-import type { IconType } from "../../../../components/Utilities/Icon"
-import type { AvatarVariant } from "../../../Information/Avatars/Avatar"
+import type { IconType } from "@/components/Utilities/Icon"
+import type { AvatarVariant } from "@/experimental/Information/Avatars/Avatar"
 
 export type SelectItemObject<T, R = unknown> = {
   type?: "item"
