+++ packages/react/src/experimental/OneDataCollection/Settings/components/SortingSelector.tsx
@@ -3,7 +3,11 @@ import { Select } from "@/experimental/Forms/Fields/Select"
 import { ArrowDown, ArrowUp, Placeholder } from "@/icons/app"
 import { useI18n } from "@/lib/providers/i18n"
 import { useEffect, useState } from "react"
-import { SortingKey, SortingsDefinition, SortingsState } from "../../sortings"
+import {
+  SortingKey,
+  SortingsDefinition,
+  SortingsState,
+} from "../../../../hooks/datasource/types/sortings.typings"
 
 export const EmptySortingValue = "__no-sorting__"
