+++ packages/react/src/experimental/OneDataCollection/index.tsx
@@ -14,6 +14,8 @@ import type {
   FiltersDefinition,
   FiltersState,
 } from "../../components/OneFilterPicker/types"
+import { SortingsDefinition } from "../../hooks/datasource/types/sortings.typings"
+import { DataError } from "../../hooks/datasource/useData"
 import { OneActionBar } from "../OneActionBar"
 import { getSecondaryActions, MAX_EXPANDED_ACTIONS } from "./actions"
 import { CollectionActions } from "./CollectionActions/CollectionActions"
@@ -22,26 +24,26 @@ import { navigationFilterTypes } from "./navigationFilters"
 import { NavigationFiltersDefinition } from "./navigationFilters/types"
 import { Search } from "./search"
 import { Settings } from "./Settings"
-import { SortingsDefinition } from "./sortings"
 import { SummariesDefinition } from "./summary"
 import type {
   BulkActionDefinition,
-  DataSource,
-  GroupingDefinition,
   OnBulkActionCallback,
   OnLoadDataCallback,
-  OnSelectItemsCallback,
-  RecordType,
 } from "./types"
-import { DataError } from "./useData"
 import { CustomEmptyStates, useEmptyState } from "./useEmptyState"
+export * from "./navigationFilters/types"
 
 import type { Visualization } from "./visualizations/collection"
 import { VisualizationRenderer } from "./visualizations/collection"
 
 const MotionIcon = motion.create(Icon)
 
-export { useDataSource } from "./useDataSource"
+import {
+  DataSource,
+  GroupingDefinition,
+  OnSelectItemsCallback,
+  RecordType,
+} from "@/hooks/datasource"
 
 /**
  * A component that renders a collection of data with filtering and visualization capabilities.
