+++ packages/react/src/experimental/OneDataCollection/property-render.ts
@@ -1,5 +1,5 @@
+import { RecordType } from "@/hooks/datasource"
 import { ReactNode } from "react"
-import { RecordType } from "./types"
 import { VisualizationType } from "./visualizations/collection/types"
 import { propertyRenderers } from "./visualizations/property"
 import { PropertyRendererMetadata } from "./visualizations/property/types"
