+++ packages/react/src/experimental/OneDataCollection/useInfiniteScrollPagination.ts
@@ -1,6 +1,5 @@
+import { isInfiniteScrollPagination, PaginationInfo } from "@/hooks/datasource"
 import { useEffect, useRef } from "react"
-import { PaginationInfo } from "../exports"
-import { isInfiniteScrollPagination } from "./useData"
 
 export const useInfiniteScrollPagination = (
   paginationInfo: PaginationInfo | null,
