+++ packages/react/src/experimental/OneDataCollection/visualizations/collection/Card/index.spec.tsx
@@ -4,12 +4,12 @@ import { render, renderHook, screen, waitFor } from "@testing-library/react"
 import { ReactNode } from "react"
 import { describe, expect, it, vi } from "vitest"
 import type { FiltersDefinition } from "../../../../../components/OneFilterPicker/types"
+import { SortingsDefinition } from "../../../../../hooks/datasource/types/sortings.typings"
+import { useData } from "../../../../../hooks/datasource/useData"
 import { I18nProvider } from "../../../../../lib/providers/i18n"
 import { defaultTranslations } from "../../../../../lib/providers/i18n/i18n-provider-defaults"
 import { ItemActionsDefinition } from "../../../item-actions"
-import { SortingsDefinition } from "../../../sortings"
 import type { DataSource, GroupingDefinition } from "../../../types"
-import { useData } from "../../../useData"
 import { CardCollection } from "./index"
 
 type Person = {
