+++ packages/react/src/experimental/OneDataCollection/visualizations/collection/Card/index.tsx
@@ -2,20 +2,18 @@ import { AvatarVariant } from "@/experimental/Information/Avatars/Avatar"
 import { OneCard } from "@/experimental/OneCard"
 import { GroupHeader } from "@/experimental/OneDataCollection/components/GroupHeader/GroupHeader"
 import { NavigationFiltersDefinition } from "@/experimental/OneDataCollection/navigationFilters/types"
-import {
-  getAnimationVariants,
-  useGroups,
-} from "@/experimental/OneDataCollection/useGroups"
-import { useSelectable } from "@/experimental/OneDataCollection/useSelectable"
+import { getAnimationVariants, useGroups } from "@/hooks/datasource/useGroups"
+import { useSelectable } from "@/hooks/datasource/useSelectable"
 import { Card, CardContent, CardHeader, CardTitle } from "@/ui/Card"
 import { Skeleton } from "@/ui/skeleton"
 import { AnimatePresence, motion } from "motion/react"
 import { useEffect, useMemo } from "react"
 import type { FiltersDefinition } from "../../../../../components/OneFilterPicker/types"
+import { SortingsDefinition } from "../../../../../hooks/datasource/types/sortings.typings"
+import { useData } from "../../../../../hooks/datasource/useData"
 import { PagesPagination } from "../../../components/PagesPagination"
 import { ItemActionsDefinition } from "../../../item-actions"
 import { PropertyDefinition, renderProperty } from "../../../property-render"
-import { SortingsDefinition } from "../../../sortings"
 import { SummariesDefinition } from "../../../summary"
 import {
   CollectionProps,
@@ -23,7 +21,6 @@ import {
   GroupingDefinition,
   RecordType,
 } from "../../../types"
-import { useData } from "../../../useData"
 
 export type CardPropertyDefinition<T> = PropertyDefinition<T>
