+++ packages/react/src/experimental/OneDataCollection/visualizations/collection/List/index.tsx
@@ -1,7 +1,7 @@
 import { NavigationFiltersDefinition } from "@/experimental/OneDataCollection/navigationFilters/types"
 
 import { GroupHeader } from "@/experimental/OneDataCollection/components/GroupHeader/GroupHeader"
-import { useGroups } from "@/experimental/OneDataCollection/useGroups"
+import { useGroups } from "@/hooks/datasource/useGroups"
 
 import { useInfiniteScrollPagination } from "@/experimental/OneDataCollection/useInfiniteScrollPagination"
 import { cn } from "@/lib/utils"
@@ -9,14 +9,17 @@ import { Skeleton } from "@/ui/skeleton"
 import { AnimatePresence, motion } from "motion/react"
 import { useEffect } from "react"
 import type { FiltersDefinition } from "../../../../../components/OneFilterPicker/types"
+import { SortingsDefinition } from "../../../../../hooks/datasource/types/sortings.typings"
+import {
+  isInfiniteScrollPagination,
+  useData,
+} from "../../../../../hooks/datasource/useData"
+import { useSelectable } from "../../../../../hooks/datasource/useSelectable"
 import { Spinner } from "../../../../Information/Spinner"
 import { PagesPagination } from "../../../components/PagesPagination"
 import { ItemActionsDefinition } from "../../../item-actions"
-import { SortingsDefinition } from "../../../sortings"
 import { SummariesDefinition } from "../../../summary"
 import { CollectionProps, GroupingDefinition, RecordType } from "../../../types"
-import { isInfiniteScrollPagination, useData } from "../../../useData"
-import { useSelectable } from "../../../useSelectable"
 import { ListGroup } from "./components/ListGroup"
 import { ListVisualizationOptions } from "./types"
