+++ packages/react/src/experimental/OneDataCollection/visualizations/collection/Table/index.tsx
@@ -2,10 +2,6 @@ import { Checkbox } from "@/experimental/Forms/Fields/Checkbox"
 import { GroupHeader } from "@/experimental/OneDataCollection/components/GroupHeader"
 import { PagesPagination } from "@/experimental/OneDataCollection/components/PagesPagination"
 import { NavigationFiltersDefinition } from "@/experimental/OneDataCollection/navigationFilters/types"
-import {
-  getAnimationVariants,
-  useGroups,
-} from "@/experimental/OneDataCollection/useGroups"
 import { useInfiniteScrollPagination } from "@/experimental/OneDataCollection/useInfiniteScrollPagination"
 import {
   OneTable,
@@ -16,23 +12,27 @@ import {
   TableHeader,
   TableRow,
 } from "@/experimental/OneTable"
+import { getAnimationVariants, useGroups } from "@/hooks/datasource/useGroups"
 import { useI18n } from "@/lib/providers/i18n"
 import { cn } from "@/lib/utils"
 import { Skeleton } from "@/ui/skeleton.tsx"
 import { AnimatePresence, motion } from "motion/react"
 import { ComponentProps, Fragment, useEffect, useMemo, useState } from "react"
 import type { FiltersDefinition } from "../../../../../components/OneFilterPicker/types"
-import { ItemActionsDefinition } from "../../../item-actions"
-import { PropertyDefinition } from "../../../property-render"
 import {
   SortingKey,
   SortingsDefinition,
   SortingsState,
-} from "../../../sortings"
+} from "../../../../../hooks/datasource/types/sortings.typings"
+import {
+  isInfiniteScrollPagination,
+  useData,
+} from "../../../../../hooks/datasource/useData"
+import { useSelectable } from "../../../../../hooks/datasource/useSelectable"
+import { ItemActionsDefinition } from "../../../item-actions"
+import { PropertyDefinition } from "../../../property-render"
 import { SummariesDefinition, SummaryKey } from "../../../summary"
 import { CollectionProps, GroupingDefinition, RecordType } from "../../../types"
-import { isInfiniteScrollPagination, useData } from "../../../useData"
-import { useSelectable } from "../../../useSelectable"
 import { statusToChecked } from "../utils"
 import { Row } from "./components/Row"
