+++ packages/react/src/hooks/datasource/types/records.typings.ts
@@ -1,0 +1,5 @@
+/**
+ * Represents a record type with string keys and unknown values.
+ * This type is used to represent the data structure of a collection.
+ */
+export type RecordType = Record<string, unknown>
