+++ packages/react/src/hooks/datasource/useDataSource.ts
@@ -1,24 +1,26 @@
-import { useEffect, useMemo, useState } from "react"
-import { useDebounceValue } from "usehooks-ts"
 import {
-  FiltersDefinition,
-  FiltersState,
   ItemActionsDefinition,
   NavigationFiltersDefinition,
   NavigationFiltersState,
-} from "./exports"
-import { navigationFilterTypes } from "./navigationFilters"
+} from "@/experimental/OneDataCollection/exports"
+import { navigationFilterTypes } from "@/experimental/OneDataCollection/navigationFilters"
 import {
   CollectionSearchOptions,
+  SummariesDefinition,
+} from "@/experimental/OneDataCollection/types"
+import { useEffect, useMemo, useState } from "react"
+import { useDebounceValue } from "usehooks-ts"
+import {
   DataSource,
   DataSourceDefinition,
+  FiltersDefinition,
+  FiltersState,
   GroupingDefinition,
   GroupingState,
   PaginationType,
   RecordType,
   SortingsDefinition,
   SortingsState,
-  SummariesDefinition,
 } from "./types"
 
 /**
