+++ packages/react/src/hooks/datasource/useGroups.ts
@@ -1,5 +1,5 @@
 import { useEffect, useState } from "react"
-import { RecordType } from "./types"
+import { RecordType } from "./types/records.typings"
 import { GroupRecord } from "./useData"
 
 export const useGroups = <R extends RecordType>(
